% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{bench_columns}
\alias{bench_columns}
\title{Add table size to the benchmark}
\usage{
bench_columns(bench_obj, column_mode = "all", ...)
}
\arguments{
\item{bench_obj}{A benchmark definition created with synthetic_bench()}

\item{column_mode}{Mode to use for benchmarking columns. The default 'all' just used the complete
table, 'single' benchmarks each column separately and 'type' separates the columns in groups of
a single type (integer, double, etc.)}

\item{...}{One or more numerical vectors defining the number of columns to use in the benchmarks.
For values larger than the number of columns in the dataset, columns are recycled using the
appropriate column definition (generator).}
}
\value{
An updated benchmark definition object
}
\description{
Add table size to the benchmark
}
