% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{bench_compression}
\alias{bench_compression}
\title{Add compression to the benchmark}
\usage{
bench_compression(bench_obj, ...)
}
\arguments{
\item{bench_obj}{A benchmark definition created with synthetic_bench()}

\item{...}{One or more numerical vectors defining the compression percentage to use for benchmarking}
}
\value{
An updated benchmark definition object
}
\description{
Add compression to the benchmark
}
