% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{bench_generators}
\alias{bench_generators}
\title{Add dataset generators to the benchmark}
\usage{
bench_generators(bench_obj, ...)
}
\arguments{
\item{bench_obj}{A benchmark definition created with synthetic_bench()}

\item{...}{One or more generators created with table_generator()}
}
\value{
An updated benchmark definition object
}
\description{
Add dataset generators to the benchmark
}
