% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{bench_rows}
\alias{bench_rows}
\title{Add table size to the benchmark}
\usage{
bench_rows(bench_obj, ...)
}
\arguments{
\item{bench_obj}{A benchmark definition created with synthetic_bench()}

\item{...}{One or more numerical vectors defining the number of rows to use for benchmarking}
}
\value{
An updated benchmark definition object
}
\description{
Add table size to the benchmark
}
