% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{bench_streamers}
\alias{bench_streamers}
\title{Add streamers to the benchmark}
\usage{
bench_streamers(bench_obj, ...)
}
\arguments{
\item{bench_obj}{A benchmark definition created with synthetic_bench()}

\item{...}{One or more table streamers created with table_streamer()}
}
\value{
An updated benchmark definition object
}
\description{
Add streamers to the benchmark
}
