% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{bench_tables}
\alias{bench_tables}
\title{Add datasets or dataset generators to the benchmark}
\usage{
bench_tables(bench_obj, ...)
}
\arguments{
\item{bench_obj}{A benchmark definition created with synthetic_bench()}

\item{...}{One or more synthetic_table or data.frame objects. A synthetic_table can be created
using method synthetic_table().}
}
\value{
An updated benchmark definition object
}
\description{
Add datasets or dataset generators to the benchmark
}
