% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{bench_threads}
\alias{bench_threads}
\title{Add thread count to the benchmark}
\usage{
bench_threads(bench_obj, ...)
}
\arguments{
\item{bench_obj}{A benchmark definition created with synthetic_bench()}

\item{...}{One or more numerical vectors defining the number of threads to use for benchmarking}
}
\value{
An updated benchmark definition object
}
\description{
Add thread count to the benchmark
}
