% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_02a_claim_covariates.R
\name{check_relativity}
\alias{check_relativity}
\title{Function to check the input Covariate Relativities}
\usage{
check_relativity(factors, relativity)
}
\arguments{
\item{factors}{named list of vectors, containing the name of the covariates
and associated factors in vector form.}

\item{relativity}{relativity dataframe as defined in \code{\link{relativity_template}}.}
}
\description{
Performs assertions on inputted relativities, will raise an error if any checks
fail. Currently checks on the: necessary column names used in dataframe,
required factors and levels based on inputted \code{factors}, inputted relativities
being non-negative numbers.
}
