% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_02a_claim_covariates.R
\name{claim_size_adj.fit}
\alias{claim_size_adj.fit}
\title{Covariates Claim Size Adjustment}
\usage{
claim_size_adj.fit(covariates_data, claim_size)
}
\arguments{
\item{covariates_data}{a \code{covariates_data} object}

\item{claim_size}{a list in the same output as \code{\link{claim_size}}}
}
\description{
Adjusts claim sizes given the covariates related to each claim. The relative adjustment of each claim size is given by the severity relativities of the covariates.
}
