% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claims_class.R
\name{claims}
\alias{claims}
\title{Construction of a \code{claims} Object}
\usage{
claims(
  frequency_vector,
  occurrence_list,
  claim_size_list,
  notification_list,
  settlement_list,
  no_payments_list,
  payment_size_list,
  payment_delay_list,
  payment_time_list,
  payment_inflated_list
)
}
\arguments{
\item{frequency_vector}{a vector of claim frequencies for all the periods.}

\item{occurrence_list}{list of claim occurrence times.}

\item{claim_size_list}{list of claim sizes.}

\item{notification_list}{list of notification delays.}

\item{settlement_list}{list of settlement delays.}

\item{no_payments_list}{list of number of partial payments.}

\item{payment_size_list}{(compound) list of payment size pattern (without
inflation).}

\item{payment_delay_list}{(compound) list of inter partial delays.}

\item{payment_time_list}{(compound) list of payment times on a \strong{continuous}
time scale.}

\item{payment_inflated_list}{(compound) list of payment size pattern (after
inflation).}
}
\value{
Returns a \code{claims} object (mainly a reformat of the arguments as a
list object), with the 10 components as listed above.
}
\description{
Contructs a \code{claims} object which stores all the simulated quantities.
}
