% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{collect.benchmark_definition}
\alias{collect.benchmark_definition}
\title{Compute generic function to start computing a benchmark}
\usage{
\method{collect}{benchmark_definition}(x, ...)
}
\arguments{
\item{x}{A benchmark definition created with synthetic_bench()}

\item{...}{}
}
\value{
Benchmark resuls
}
\description{
Compute generic function to start computing a benchmark
}
