% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{compute.benchmark_definition}
\alias{compute.benchmark_definition}
\title{Compute generic function to start computing a benchmark}
\usage{
compute.benchmark_definition(bench_obj)
}
\arguments{
\item{bench_obj}{A benchmark definition created with synthetic_bench()}
}
\value{
Benchmark resuls
}
\description{
Compute generic function to start computing a benchmark
}
