% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{construct_synth_bench}
\alias{construct_synth_bench}
\title{Constructs a synth_bench class}
\usage{
construct_synth_bench(nr_of_runs = 10, cycle_size = 10,
  result_folder = "results", progress = TRUE)
}
\arguments{
\item{nr_of_runs}{repeat the benchmark for statistics}

\item{cycle_size}{create cycly_size files before overwriting}

\item{result_folder}{folder to use for temporal storage of results}

\item{progress}{if TRUE, a progress bar is displayed with the progress and estimated
remaining time for the benchmark}
}
\value{
A synth_bench class
}
\description{
Constructs a synth_bench class
}
