% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_02a_claim_covariates.R
\name{covariates_data}
\alias{covariates_data}
\title{Construction of a \code{covariates_data} Object}
\usage{
covariates_data(covariates, data, covariates_id = NULL)
}
\arguments{
\item{covariates}{a \code{\link{covariates}} object}

\item{data}{a dataset of covariate values, with columns equal to each of the covariate factors and rows related to individual claim observations.}

\item{covariates_id}{optional list of list of ids, in the same format as a \code{\link{claim_size}} output. Also see \code{\link{to_SynthETIC}}. Defaults to \code{NULL}.}
}
\value{
Returns a \code{covariates_data} object.
}
\description{
Constructs a \code{covariates_data} object which stores the dataset of known covariate levels of each factor.
}
