% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_02a_claim_covariates.R
\name{covariates_relativity}
\alias{covariates_relativity}
\title{Calculates Relativities}
\usage{
covariates_relativity(
  covariates_data,
  freq_sev = c("freq", "sev"),
  by_ids = FALSE
)
}
\arguments{
\item{covariates_data}{a \code{\link{covariates_data}} object}

\item{freq_sev}{one of \code{freq} or \code{sev} to calculate the frequency or severity relativity respectively.}

\item{by_ids}{optional boolean to calculate reorder the result based off claim observations instead of observations in the covariates dataset. Defaults to FALSE.}
}
\description{
Calculates the relativities (\code{freq} or \code{sev}) of a set of covariate values.
}
