% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds.}
}
\description{
Returns the observed coefficient of variation (CoV) of a given sample
\code{x}. \cr \cr
If \code{na.rm} is true then missing values are removed before
computation proceeds, as in the case of the \code{mean()} function.
}
\details{
The coefficient of variation is defined as is defined as the
ratio of the standard deviation to the mean. It shows the extent of
variability in relation to the mean of the population. \cr \cr
\code{cv()} estimates the CoV of a given sample by computing the ratio of
the sample standard deviation (see \code{stats::sd}) to the sample mean.
}
\examples{
cv(1:10)
}
