% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_numerical.R
\name{dbl_template_from_column}
\alias{dbl_template_from_column}
\title{Create a numerical column template from a source vector}
\usage{
dbl_template_from_column(column)
}
\arguments{
\item{column}{numerical vector}
}
\value{
a numerical column template
}
\description{
Create a numerical column template from a source vector
}
