% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed.R
\name{delayed_eval}
\alias{delayed_eval}
\title{Evaluate a delayed expression}
\usage{
delayed_eval(delayed_expr)
}
\arguments{
\item{delayed_expr}{delayed expression generated with method `delayed_expr()``}
}
\value{
the evaluated expression
}
\description{
Evaluate a delayed expression
}
\examples{
# define a delayed expression
x <- delayed_expr(size * 5 + 1)

# create a new expression
y = 1 + x + 5 * x - x

\dontrun{
# generates an error, as size is not defined
delayed_eval(y)
}

# calculates expression y using size
size = 0.1
delayed_eval(y)
}
