% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed.R
\name{delayed_expr}
\alias{delayed_expr}
\title{A wrapped expression that can be used with operators to generate new expressions}
\usage{
delayed_expr(expr, expr_str)
}
\arguments{
\item{expr}{an expression}
}
\value{
A delayed_expr object
}
\description{
A wrapped expression that can be used with operators to generate new expressions
}
\examples{
# define a delayed expression
f_delay <- function() {
  delayed_expr(size)
}

# create a new expression
y = (1 + f_delay()) == 5

# generates an error, as size is not defined
delayed_eval(y)

# calculates expression y using size
size = 0.1
delayed_eval(y)
}
