% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed.R
\name{delayed_to_str}
\alias{delayed_to_str}
\title{Textual representation of a delayed expression}
\usage{
delayed_to_str(delayed_expr)
}
\arguments{
\item{delayed_expr}{delayed expression such as that generated with nr_of_rows()}
}
\value{
well formatted string representation of the delayed expression
}
\description{
This is a convenience method to use when the delayed expression needs to be converted to
a string, for example in the print function of a vector template that allows for methods
like nr_of_rows().
}
\examples{
delayed_to_str(nr_of_rows())
}
