% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{generate_transaction_dataset}
\alias{generate_transaction_dataset}
\title{Generate a Transactions Dataset}
\usage{
generate_transaction_dataset(claims, adjust = FALSE)
}
\arguments{
\item{claims}{an \code{claims} object containing all the simulated quantities,
see \code{\link{claims}}.}

\item{adjust}{if TRUE then the payment times will be forced to match with the
maximum development period under consideration; default FALSE (which will
produce out-of-bound payment times).}
}
\value{
A dataframe that takes the same structure as
\code{\link{test_transaction_dataset}}.
}
\description{
Generates a dataset of partial payment records that takes the same structure
as \code{test_transaction_dataset} included in this package, with each row
representing a unique payment.
}
\examples{
# this generates the built-in test_transaction_dataset
transact_data <- generate_transaction_dataset(test_claims_object)

}
\seealso{
\code{\link{test_transaction_dataset}}
}
