% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_Beta_parameters}
\alias{get_Beta_parameters}
\title{Estimating Beta Parameters}
\usage{
get_Beta_parameters(target_mean, target_cv)
}
\arguments{
\item{target_mean}{mean of the target Beta distribution (between 0 and 1).}

\item{target_cv}{CoV of the target Beta distribution.}
}
\description{
Returns the Beta parameters given the mean and the CoV of the target Beta
distribution.
}
\examples{
get_Beta_parameters(target_mean = 0.5, target_cv = 0.20)
get_Beta_parameters(target_mean = 0.5,
                    target_cv = c(0.10, 0.20, 0.30))
}
