% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_template.R
\name{nr_of_rows}
\alias{nr_of_rows}
\title{Number of rows drop-in}
\usage{
nr_of_rows()
}
\value{
a delayed expression for the number of rows
}
\description{
When defining a template, you can use this method as a drop-in for the unknown number
of rows that the user selects
}
\examples{
template_integer(1, nr_of_rows())
}
