% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed.R
\name{or}
\alias{or}
\alias{||.delayed_expr}
\title{Or operator override for delayed expressions}
\usage{
`||.delayed_expr`(x, y)
}
\arguments{
\item{x}{delayed expression or object}

\item{y}{delayed expression or object}
}
\description{
Or operator override for delayed expressions
}
