% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plot.claims}
\alias{plot.claims}
\title{Plot of Cumulative Claims Payments (Incurred Pattern)}
\usage{
\method{plot}{claims}(x, by_year = FALSE, inflated = TRUE, adjust = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{claims} containing all the simulated quantities.}

\item{by_year}{if \code{TRUE} returns a plot by occurrence year; otherwise returns
a plot by occurrence period (default).}

\item{inflated}{if \code{TRUE} shows a plot of payment pattern after inflation;
otherwise shows a plot of discounted payment pattern.}

\item{adjust}{if \code{TRUE} then the payment times will be forced to match with
the maximum development period under consideration, otherwise the plot will
see claims beyond the maximum development period; default \code{TRUE}.}

\item{...}{other graphical parameters.}
}
\description{
Generates a plot of cumulative claims paid (as a percentage of total amount
incurred) as a function of development time for each occurrence period.
}
\examples{
plot(test_claims_object)
plot(test_claims_object, adjust = FALSE)
}
\seealso{
\code{\link{claims}}
}
