% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{print.benchmark_definition}
\alias{print.benchmark_definition}
\title{Print details of synthetic benchmark}
\usage{
\method{print}{benchmark_definition}(x, ...)
}
\arguments{
\item{x}{synt_bench object}

\item{...}{other arguments to print}
}
\description{
Print details of synthetic benchmark
}
