% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_generator.R
\name{print.vectortemplate}
\alias{print.vectortemplate}
\title{Print a vector template description}
\usage{
\method{print}{vectortemplate}(x, ...)
}
\arguments{
\item{x}{vector template}

\item{...}{additional parameters (will be discarded)}
}
\description{
Print a vector template description
}
