% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{return_parameters}
\alias{return_parameters}
\title{Get Current Parameters}
\usage{
return_parameters(print = FALSE)
}
\arguments{
\item{print}{logical; if TRUE prints a message.}
}
\description{
Returns the current values of \code{ref_claim} and \code{time_unit} parameters, two
packagewise-global variables used by all simulation functions within this
package.
}
\details{
Returns and (optionally) prints the current values of \code{ref_claim}
and \code{time_unit} parameters.
}
\examples{
cur <- return_parameters()
cur
set_parameters(ref_claim = 200000, time_unit = 1/12) # monthly reserving
return_parameters(print = FALSE)
}
\seealso{
\code{\link{set_parameters}}
}
