% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspline.R
\name{rspline}
\alias{rspline}
\title{Random number generator with custom distribution}
\usage{
rspline(n, control_values, seed)
}
\arguments{
\item{n}{number of observations}

\item{control_values}{values describing the ordered distribution}

\item{seed}{single numeric to specify the seed}
}
\value{
a vector of random numbers generated from the spline interpolated custom distribution.
}
\description{
Random numbers are taken from a distribution which is defined by control points.
}
