% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_logical.R
\name{sample_logical}
\alias{sample_logical}
\title{Generate a logical vector with certain distribution of TRUE, FALSE and NA}
\usage{
sample_logical(length, true_false_na_ratio = c(1, 1, 0))
}
\arguments{
\item{length}{length of the vector}

\item{true_false_na_ratio}{relative ratio of values TRUE, FALSE and NA}
}
\value{
logical vector
}
\description{
Generate a logical vector with certain distribution of TRUE, FALSE and NA
}
