% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_numerical.R
\name{sample_numerical_normal}
\alias{sample_numerical_normal}
\title{Generate a numerical vector with normal distribution}
\usage{
sample_numerical_normal(length, mean = 0, sd = 1,
  max_distict_values = NULL)
}
\arguments{
\item{length}{length of the vector}

\item{mean}{mean of the distribution}

\item{sd}{standard deviation of the distribution}

\item{max_distict_values}{maximum number of disctict values in the vector}
}
\value{
numerical vector
}
\description{
Generate a numerical vector with normal distribution
}
