% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_numerical.R
\name{sample_numerical_uniform}
\alias{sample_numerical_uniform}
\title{Generate a numerical vector with uniform distribution}
\usage{
sample_numerical_uniform(length, min_value = 0, max_value = 1,
  max_distict_values = NULL)
}
\arguments{
\item{length}{length of the vector}

\item{min_value}{minimum value in the vector}

\item{max_value}{maximum value in the vector}

\item{max_distict_values}{maximum number of disctict values in the vector}
}
\value{
numerical vector
}
\description{
Generate a numerical vector with uniform distribution
}
