% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_string.R
\name{sample_string}
\alias{sample_string}
\title{Generate a character vector with certain distribution of string lengths}
\usage{
sample_string(length, min_str_size = 1, max_str_size = 10,
  max_distict_values = NULL)
}
\arguments{
\item{length}{length of the vector}

\item{min_str_size}{minimum string length}

\item{max_str_size}{maximum string length}

\item{max_distict_values}{maximum number of disctict values in the vector}
}
\value{
character vector
}
\description{
Generate a character vector with certain distribution of string lengths
}
