% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_02a_claim_covariates.R
\name{set.covariates_relativity}
\alias{set.covariates_relativity}
\title{Sets the claims relativity for a \code{covariates} object.}
\usage{
set.covariates_relativity(covariates, relativity, freq_sev = c("freq", "sev"))
}
\arguments{
\item{covariates}{an object of type \code{covariates}, see \code{\link{covariates}}}

\item{relativity}{see \code{\link{relativity_template}}}

\item{freq_sev}{one of \code{"freq"} or \code{"sev"} to adjust frequency or severity
relativities respectively}
}
\value{
Returns a \code{covariates} object.
}
\description{
Sets the claims relativity for a \code{covariates} object.
}
\seealso{
\code{\link{covariates}}, \code{\link{relativity_template}}
}
