% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_02a_claim_covariates.R
\name{set.covariates_relativity_F}
\alias{set.covariates_relativity_F}
\title{Sets the claim frequency relativity for a \code{covariates} object.}
\usage{
set.covariates_relativity_F(covariates, relativity)
}
\arguments{
\item{covariates}{an object of type \code{covariates}. see \code{\link{covariates}}}

\item{relativity}{see \code{\link{relativity_template}}}
}
\value{
Returns a \code{covariates} object.
}
\description{
Sets the claim frequency relativity for a \code{covariates} object.
}
