% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_02a_claim_covariates.R
\name{set.covariates_relativity_R}
\alias{set.covariates_relativity_R}
\title{Sets the claim size relativity for a \code{covariates} object.}
\usage{
set.covariates_relativity_R(covariates, relativity)
}
\arguments{
\item{covariates}{an object of type \code{covariates}. See \code{\link{covariates}}}

\item{relativity}{see \code{\link{relativity_template}}}
}
\value{
Returns a \code{covariates} object.
}
\description{
Sets the claim size relativity for a \code{covariates} object.
}
