% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_distribution.R
\name{sspline}
\alias{sspline}
\title{Random number generator with custom distribution}
\usage{
sspline(control_points, values)
}
\arguments{
\item{control_points}{values describing the ordered distribution}

\item{values}{values in the distribution to simulate}
}
\value{
a vector of numbers generated from the spline interpolated custom distribution.
}
\description{
Random numbers are taken from a distribution which is defined by control points.
}
