% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamers.R
\name{streamer_arrow}
\alias{streamer_arrow}
\title{arrow streamer}
\usage{
streamer_arrow(id = "arrow")
}
\arguments{
\item{id}{Identifier to mark the streamer in benchmarks (default: 'arrow')}
}
\value{
A table streamer that uses the arrow format
}
\description{
arrow streamer
}
