% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamers.R
\name{streamer_datatable}
\alias{streamer_datatable}
\title{fread/fwrite streamer}
\usage{
streamer_datatable(id = "data.table")
}
\arguments{
\item{id}{Identifier to mark the streamer in benchmarks (default: 'data.table')}
}
\value{
A table streamer that uses data.table's fread/fwrite to read and write the csv format
}
\description{
fread/fwrite streamer
}
