% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamers.R
\name{streamer_feather}
\alias{streamer_feather}
\title{feather streamer}
\usage{
streamer_feather(id = "feather")
}
\arguments{
\item{id}{Identifier to mark the streamer in benchmarks (default: 'feather')}
}
\value{
A table streamer that uses the feather format
}
\description{
feather streamer
}
