% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamers.R
\name{streamer_fst}
\alias{streamer_fst}
\title{fst streamer}
\usage{
streamer_fst(id = "fst")
}
\arguments{
\item{id}{Identifier to mark the streamer in benchmarks (default: 'fst')}
}
\value{
A table streamer that uses the fst format
}
\description{
fst streamer
}
