% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamers.R
\name{streamer_parguet}
\alias{streamer_parguet}
\title{parguet streamer}
\usage{
streamer_parguet(id = "parguet")
}
\arguments{
\item{id}{Identifier to mark the streamer in benchmarks (default: 'parguet')}
}
\value{
A table streamer that uses the parguet format
}
\description{
parguet streamer
}
