% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamers.R
\name{streamer_qs}
\alias{streamer_qs}
\title{qs streamer}
\usage{
streamer_qs(id = "qs", preset = "high")
}
\arguments{
\item{id}{Identifier to mark the streamer in benchmarks (default: 'qs')}

\item{preset}{Use one of the preset compression settings}
}
\value{
A table streamer that uses the fst format
}
\description{
qs streamer
}
