% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamers.R
\name{streamer_rds}
\alias{streamer_rds}
\title{baseR streamer}
\usage{
streamer_rds(id = "rds")
}
\arguments{
\item{id}{Identifier to mark the streamer in benchmarks (default: 'rds')}
}
\value{
A table streamer that uses the rds format
}
\description{
baseR streamer
}
