% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{synthetic_bench}
\alias{synthetic_bench}
\title{Runs benchmarks}
\usage{
synthetic_bench(bench_id, table_generator, table_streamer, nr_of_rows,
  compression, nr_of_runs = 100, cycle_size = 10,
  result_folder = "results")
}
\arguments{
\item{table_generator}{function f(nr_of_rows) that generates the data.frame}

\item{table_streamer}{object of class tablestreamer that defines the streaming
methods, generated with table_streamer().}

\item{nr_of_rows}{number of rows to use in the benchmark}

\item{compression}{vector of compression values to use for benchmarking}

\item{nr_of_runs}{repeat the benchmark for statistics}

\item{cycle_size}{create cycly_size files before overwriting}

\item{result_folder}{folder to use for temporal storage of results}

\item{id}{benchmark id (e.g. 'fst', 'rds', 'arrow')}

\item{table_reader}{function f(file_name) that reads a file into a data.frame}
}
\value{
benchmarks results
}
\description{
Runs benchmarks
}
