% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_definition.R
\name{table_definition}
\alias{table_definition}
\title{Define an object of class 'tablegenerator' that defines specific methods for
creation of a dataset}
\usage{
table_definition(id, ...)
}
\arguments{
\item{id}{ID of the dataset (e.g. 'fst homepage dataset')}

\item{...}{}
}
\value{
a table definition that can be used to generate synthetic data using generate()
}
\description{
Define an object of class 'tablegenerator' that defines specific methods for
creation of a dataset
}
