% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_generator.R
\name{table_generator}
\alias{table_generator}
\title{Define an object of class 'tablegenerator' that defines specific methods for
creation of a dataset}
\usage{
table_generator(id, generator)
}
\arguments{
\item{id}{ID of the table generator (e.g. 'single column sparse')}

\item{generator}{method with signature f(nr_of_rows) that generates a data.frame}
}
\value{
a tablestreamer object
}
\description{
Define an object of class 'tablegenerator' that defines specific methods for
creation of a dataset
}
