% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_double.R
\name{template_double_from_vec}
\alias{template_double_from_vec}
\title{Create a numerical column template from a source vector}
\usage{
template_double_from_vec(vec)
}
\arguments{
\item{vec}{vector to estimate the template from}
}
\value{
a numerical column template
}
\description{
Create a numerical column template from a source vector
}
