% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_integer.R
\name{template_integer}
\alias{template_integer}
\title{Define a template for integer vectors with custom distribution and characteristics}
\usage{
template_integer(min_value = 1 - .Machine$integer.max,
  max_value = .Machine$integer.max, max_distict_values = NULL)
}
\arguments{
\item{min_value}{minimum value in the vector}

\item{max_value}{maximum value in the vector}

\item{max_distict_values}{maximum number of disctict values in the vector}

\item{length}{length of the vector}
}
\value{
integer vector
}
\description{
Define a template for integer vectors with custom distribution and characteristics
}
