% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_logical.R
\name{template_logical}
\alias{template_logical}
\title{Define a template for logical vectors with custom distribution of TRUE, FALSE and NA}
\usage{
template_logical(true_false_na_ratio = c(1, 1, 0))
}
\arguments{
\item{true_false_na_ratio}{relative ratio of values TRUE, FALSE and NA}
}
\value{
logical vector
}
\description{
Define a template for logical vectors with custom distribution of TRUE, FALSE and NA
}
