% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_numerical.R
\name{template_numerical_normal}
\alias{template_numerical_normal}
\title{Generate a numerical vector with normal distribution}
\usage{
template_numerical_normal(mean = 0, sd = 1,
  max_distict_values = NULL)
}
\arguments{
\item{mean}{mean of the distribution}

\item{sd}{standard deviation of the distribution}

\item{max_distict_values}{maximum number of disctict values in the vector}
}
\value{
numerical vector
}
\description{
Generate a numerical vector with normal distribution
}
