% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_raw.R
\name{template_raw}
\alias{template_raw}
\title{Generate a raw vector with custom distribution}
\usage{
template_raw(min_value = 0, max_value = 255, max_distict_values = NULL)
}
\arguments{
\item{min_value}{minimum value in the returned raw vector}

\item{max_value}{maximum value in the returned raw vector}

\item{max_distict_values}{maximum number of disctict values in the vector}
}
\value{
raw vector
}
\description{
Generate a raw vector with custom distribution
}
