% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_string.R
\name{template_string_random}
\alias{template_string_random}
\title{Generate a character vector with certain distribution of string lengths}
\usage{
template_string_random(
  min_str_size = 1,
  max_str_size = 10,
  max_distict_values = NULL
)
}
\arguments{
\item{min_str_size}{minimum string length}

\item{max_str_size}{maximum string length}

\item{max_distict_values}{maximum number of disctict values in the vector}
}
\value{
character vector
}
\description{
Generate a character vector with certain distribution of string lengths
}
