% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_claim_dataset}
\alias{test_claim_dataset}
\title{Claims Dataset}
\format{
A data frame with 3,624 rows and 7 variables:
\describe{
\item{claim_no}{claim number, which uniquely characterises each claim.}
\item{occurrence_period}{integer; period of ocurrence of the claim.}
\item{occurrence_time}{double; time of occurrence of the claim.}
\item{claim_size}{size of the claim (in constant dollar values).}
\item{notidel}{notification delay of the claim, i.e. time from occurrence to notification.}
\item{setldel}{settlement delay of the claim, i.e. time from notification to settlement.}
\item{no_payment}{number of partial payments required for the claim.}
}
}
\usage{
test_claim_dataset
}
\description{
A dataset of 3,624 rows containing individual claims features.
}
\examples{
# see a distribution of payment counts
table(test_claim_dataset$no_payment)
}
\keyword{datasets}
