% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_claim_dataset_cov}
\alias{test_claim_dataset_cov}
\title{Claims Dataset}
\format{
An object of class \code{data.frame} with 3624 rows and 7 columns.
}
\usage{
test_claim_dataset_cov
}
\description{
The \code{test_claim_dataset} where the default set of covariates have been applied to adjust claim sizes.
}
\examples{
# see a distribution of payment counts
table(test_claim_dataset_cov$no_payment)
}
\keyword{datasets}
