% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_claims_object_cov}
\alias{test_claims_object_cov}
\title{Claims Data in List Format}
\format{
An object of class \code{claims} of length 10.
}
\usage{
test_claims_object_cov
}
\description{
The \code{test_claims_object} where the default set of covariates have been applied to adjust claim sizes.
}
\examples{
test_claims_object$frequency_vector
}
\keyword{datasets}
