% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_covariates_dataset}
\alias{test_covariates_dataset}
\title{Covariates Data Object}
\format{
A \code{covariates_data} object with 3 components:
\describe{
\item{data}{data.frame; a dataset of covariate levels}
\item{covariates}{covariates; a \code{covariates} object}
\item{ids}{list; indices of the covariate-level dataset for each claim}
}
}
\usage{
test_covariates_dataset
}
\description{
An object detailing the set of covariates for each claim in the default setting of \code{SynthETIC}.
}
\examples{
test_covariates_dataset$data
}
\keyword{datasets}
