% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_covariates_obj}
\alias{test_covariates_obj}
\title{Covariates Object}
\format{
A \code{covariates} object with 3 components:
\describe{
\item{factors}{list; levels within each factor.}
\item{relativity_freq}{data.frame; first and second order frequency relativities between all the levels of each factor}
\item{relativity_sev}{data.frame; first and second order severity relativities between all the levels of each factor}
}
}
\usage{
test_covariates_obj
}
\description{
A list containing the frequency and severity relativities for three factors.
}
\examples{
test_covariates_obj$factors
}
\keyword{datasets}
