% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_transaction_dataset}
\alias{test_transaction_dataset}
\title{Transactions Dataset}
\format{
A data frame with 18,983 rows and 12 variables:
\describe{
\item{claim_no}{claim number, which uniquely characterises each claim.}
\item{pmt_no}{payment number, identification number of partial payments in respect of a particular \code{claim_no}.}
\item{occurrence_period}{integer; period of ocurrence of the claim.}
\item{occurrence_time}{double; time of occurrence of the claim.}
\item{claim_size}{size of the claim (in constant dollar values).}
\item{notidel}{notification delay of the claim, i.e. time from occurrence to notification.}
\item{setldel}{settlement delay of the claim, i.e. time from notification to settlement.}
\item{payment_time}{double; time of payment (on a continuous time scale).}
\item{payment_period}{integer; time of payment (in calendar period).}
\item{payment_size}{size of the payment in constant dollar terms.}
\item{payment_inflated}{actual size of the payment (i.e. with inflation).}
\item{payment_delay}{inter partial delay associated with the payment.}
}
}
\usage{
test_transaction_dataset
}
\description{
A dataset of 18,983 records of partial payments associated with the 3,624
claims in \code{test_claim_dataset}.
}
\keyword{datasets}
