% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_transaction_dataset_cov}
\alias{test_transaction_dataset_cov}
\title{Transactions Dataset}
\format{
An object of class \code{data.frame} with 17382 rows and 12 columns.
}
\usage{
test_transaction_dataset_cov
}
\description{
The \code{test_transaction_dataset} where the default set of covariates have been applied to adjust claim sizes.
}
\keyword{datasets}
