% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_generator.R
\name{vector_template}
\alias{vector_template}
\title{Create a template for creation of a vector with custom distribution}
\usage{
vector_template(metadata, generator, print_method = NULL)
}
\arguments{
\item{metadata}{data needed by the generator function to create a synthetic vector}

\item{generator}{function with signature f(metadata, length) that generates a
vector from the metadata and the size}

\item{print_method}{function with signature f(metadata) that prints the vector
template description. Will be used when  print() is called.}
}
\value{
an object of class 'columntemplate' that can be used with generate() to generate
vectors according to the characteristics stored in metadata
}
\description{
Create a template for creation of a vector with custom distribution
}
