% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_02a_claim_covariates.R
\name{claim_size_adj}
\alias{claim_size_adj}
\title{Covariates Claim Size Adjustment}
\usage{
claim_size_adj(covariate_obj, claim_size)
}
\arguments{
\item{covariate_obj}{a \code{covariates} object}

\item{claim_size}{a list in the same output as \code{\link{claim_size}}}
}
\description{
Adjusts claim sizes given a set of covariates. Note that this function firstly simulates covariate levels for each claim, see \code{\link{simulate_covariates}}.
}
