% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_02a_claim_covariates.R
\name{simulate_covariates}
\alias{simulate_covariates}
\title{Covariates Simulation}
\usage{
simulate_covariates(
  covariates,
  frequency_vector = 1,
  claim_size_list = list(1)
)
}
\arguments{
\item{covariates}{a \code{covariates} object}

\item{frequency_vector}{a vector in the same output as \code{\link{claim_frequency}}.}

\item{claim_size_list}{optional if \code{frequency_vector} is not inputted. A list in the same output as \code{\link{claim_size}}.}
}
\description{
Simulates covariates for each claim. The relative occurrence of each combination of covariates is given the frequency relativities of the covariates.
}
