% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamers.R
\name{streamer_vroom}
\alias{streamer_vroom}
\title{vroom streamer}
\usage{
streamer_vroom(id = "vroom")
}
\arguments{
\item{id}{Identifier to mark the streamer in benchmarks (default: 'vroom')}
}
\value{
A table streamer that uses vroom to read and write the csv format
}
\description{
vroom streamer
}
