% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthETIC-package.R
\docType{package}
\name{SynthETIC-package}
\alias{SynthETIC}
\alias{SynthETIC-package}
\title{SynthETIC: Synthetic Experience Tracking Insurance Claims}
\description{
Creation of an individual claims simulator which generates various
    features of non-life insurance claims. An initial set of test parameters,
    designed to mirror the experience of an Auto Liability portfolio, were set
    up and applied by default to generate a realistic test data set of
    individual claims (see vignette). The simulated data set then allows
    practitioners to back-test the validity of various reserving models and to
    prove and/or disprove certain actuarial assumptions made in claims
    modelling. The distributional assumptions used to generate this data set can
    be easily modified by users to match their experiences. Reference: Avanzi B,
    Taylor G, Wang M, Wong B (2020) "SynthETIC: an individual insurance claim
    simulator with feature control" <arXiv:2008.05693>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agi-lab/SynthETIC}
  \item Report bugs at \url{https://github.com/agi-lab/SynthETIC/issues}
}

}
\author{
\strong{Maintainer}: Melantha Wang \email{wang.melantha@gmail.com}

Authors:
\itemize{
  \item Benjamin Avanzi \email{b.avanzi@unimelb.edu.au}
  \item Greg Taylor \email{greg_taylor60@hotmail.com}
  \item Bernard Wong \email{bernard.wong@unsw.edu.au}
}

}
\keyword{internal}
