% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_bench.R
\name{synthetic_bench}
\alias{synthetic_bench}
\title{Define a benchmark}
\usage{
synthetic_bench(
  nr_of_runs = 3,
  cycle_size = 3,
  result_folder = "results",
  progress = TRUE
)
}
\arguments{
\item{nr_of_runs}{repeat the benchmark for statistics}

\item{cycle_size}{create cycle_size files before overwriting}

\item{result_folder}{folder to use for temporal storage of results}

\item{progress}{if TRUE, a progress bar is displayed with the progress and estimated
remaining time for the benchmark}
}
\value{
A synth_bench class that can be used in pipes
}
\description{
Define a benchmark
}
