% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_table.R
\name{synthetic_table}
\alias{synthetic_table}
\title{Create the blueprints of a synthetic table}
\usage{
synthetic_table(..., id = "synthetic table")
}
\arguments{
\item{...}{column templates created with the various template_ methods}

\item{id}{ID of the synthetic table (e.g. 'fst homepage dataset')}
}
\value{
a synthetic table template that can be used to generate synthetic data using generate()
}
\description{
Create the blueprints of a synthetic table
}
